cart items rendering

+++ eCom-backend/src/main/java/com/eCommerce/backend/dto/CartItemDto.java
@@ -3,12 +3,14 @@ package com.eCommerce.backend.dto;
 import com.eCommerce.backend.model.CartItem;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemDto {
     private long id;
     private long productId;
     private String productName;
-    private double productPrice;
+    private BigDecimal productPrice;
     private int quantity;
 
     public CartItemDto(CartItem cartItem) {
